+++ packages/brick_sqlite_generators/lib/src/sqlite_serdes_generator.dart
@@ -4,6 +4,7 @@ import 'package:brick_sqlite_abstract/annotations.dart';
 import 'package:brick_sqlite_abstract/db.dart' show InsertForeignKey;
 import 'package:brick_sqlite_abstract/sqlite_model.dart';
 import 'package:source_gen/source_gen.dart';
+import 'package:meta/meta.dart';
 
 import 'sqlite_fields.dart';
 
@@ -31,16 +32,13 @@ abstract class SqliteSerdesGenerator<_Model extends SqliteModel>
   @override
   SharedChecker checkerForField(FieldElement field) => checkerForType(field.type);
 
-  /// Generate foreign key column if the type is a sibling;
-  /// otherwise, return the field's annotated name;
-  @override
-  String providerNameForField(annotatedName, {required checker}) {
-    if (checker.isSibling) {
-      return InsertForeignKey.foreignKeyColumnName(
-          checker.unFuturedType.getDisplayString(withNullability: false), annotatedName);
+  /// from dart:collections, instead of importing a whole package
+  @protected
+  T? firstWhereOrNull<T>(Iterable<T> items, bool Function(T item) test) {
+    for (var item in items) {
+      if (test(item)) return item;
     }
-
-    return annotatedName ?? '';
+    return null;
   }
 
   @override
@@ -62,6 +60,24 @@ abstract class SqliteSerdesGenerator<_Model extends SqliteModel>
 
       return false;
     }
+
+    if (doesDeserialize) {
+      if (checker.fromJsonConstructor != null) return false;
+    } else {
+      if (checker.toJsonMethod != null) return false;
+    }
     return super.ignoreCoderForField(field, annotation, checker);
+  }
+
+  /// Generate foreign key column if the type is a sibling;
+  /// otherwise, return the field's annotated name;
+  @override
+  String providerNameForField(annotatedName, {required checker}) {
+    if (checker.isSibling) {
+      return InsertForeignKey.foreignKeyColumnName(
+          checker.unFuturedType.getDisplayString(withNullability: false), annotatedName);
+    }
+
+    return annotatedName ?? '';
   }
 }