+++ packages/brick_json_generators/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Unreleased
 
+## 1.0.1
+
+* Serialize using `fromJson` or `toJson` when defined on classes of specific fields
+
 ## 1.0.0
 
 * Separate `brick_json_generators` to its own package from `brick_rest_generators`